fix: 파일구조, 주석 변경
