+++ src/api/metro/escalatorLocal.js
@@ -1,4 +1,4 @@
-// src/api/escalatorLocal.js
+// src/api/metro/escalatorLocal.js
 // Source: src/assets/metro-data/metro/escalator/서울교통공사_에스컬레이터 설치 정보_20250310.json
 
 import escJson from "../../assets/metro-data/metro/escalator/서울교통공사_에스컬레이터 설치 정보_20250310.json";
