+++ src/api/metro/lockerLocal.js
@@ -1,4 +1,4 @@
-// src/api/lockerLocal.js
+// src/api/metro/lockerLocal.js
 // Source:
 //   src/assets/metro-data/metro/lostandFound/서울교통공사_물품보관함 위치정보_20240930.json
